Invert the meaning of binding level
Using a top down level made splicing inconvient as the references
need to be re-leveled during splice. Using distance from referene
(like de Bruijn indexes are suppposed to be) instead of distance
from root makes quotes relocatable without re-leveling.

Fixed scoping of quotes and splices to be phase consistent which
the is a problem the root distance indexes hid.

Still doesn't correctly support open code sequences that are
relocated as the distance from the reference could change on splice
and closed code might close over a reference that is not captured.